Nowhere a potentially negative value is given to getLocalVarName. Therefore, change the type to size_t to get rid of some conversion warnings.
git-svn-id: https://svn.berlios.de/svnroot/repos/escript/branches/0.6.3.experimental@350 5693b1fc-3b75-4070-9b6b-3ce3692a40d5

+++ EScript/Instructions/InstructionBlock.h
@@ -47,7 +47,7 @@ class InstructionBlock {
 		int getLine(const size_t index)const						{	return index<instructions.size() ? instructions[index].getLine() : -1;	}
 
 
-		StringId getLocalVarName(const int index)const;
+		StringId getLocalVarName(const size_t index)const;
 		const std::vector<StringId> & getLocalVariables()const		{	return localVariables;	}
 
 		size_t getNumLocalVars()const								{	return localVariables.size();	}
